@@ -1,7 +1,6 @@
 {
   "name": "charts",
   "version": "0.1.0",
-  "homepage": "/",
   "private": true,
   "dependencies": {
     "@sentry/browser": "^7.119.1",
@@ -9,23 +8,24 @@
     "bootstrap": "^4.3.1",
     "font-awesome": "^4.7.0",
     "highcharts": "^9.0.0",
-    "qrcode.react": "^0.9.2",
-    "react": "^16.7.0",
+    "qrcode.react": "^3.1.0",
+    "react": "^18.0.0",
     "react-bootstrap": "^1.0.0-beta.5",
-    "react-dom": "^16.7.0",
+    "react-dom": "^18.0.0",
     "react-github-btn": "^1.0.5",
-    "react-loader-spinner": "^2.3.0",
-    "react-scripts": "5.0.1",
+    "react-loader-spinner": "^5.4.0",
     "react-toastify": "^4.5.2"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "cypress run",
-    "eject": "react-scripts eject"
+  "devDependencies": {
+    "@vitejs/plugin-react": "^4.3.0",
+    "vite": "^6.0.0"
   },
-  "eslintConfig": {
-    "extends": "react-app"
+  "scripts": {
+    "start": "vite",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "cypress run"
   },
   "browserslist": [
     ">0.2%",
