@@ -2,7 +2,7 @@ import axios from "axios";
 import { toast } from "react-toastify";
 import log from "./logService";
 
-axios.defaults.baseURL = process.env.REACT_APP_API_URL;
+axios.defaults.baseURL = import.meta.env.VITE_API_URL;
 
 axios.interceptors.response.use(null, error => {
   const expectedError =
